+++ Tools/mkzip.py
@@ -1,9 +1,7 @@
-# C:/Users/aathl/AppData/Local/Microsoft/WindowsApps/python.exe
-
 # MSC Music Manager Packing Zip
 # This script let's you quickly compress new release to .zip file.
 # It also removes logs and history from folder.
-# Script version: 2.0 (27.09.2019)
+# Script version: 2.1.1 (14.10.2019)
 #
 # This file is distributed under the same license as the MSCMM is.
 
@@ -14,12 +12,24 @@ from zipfile import ZipFile
 from array import array
 import shutil
 
-print("=== MSCMM Packing Script 2.0 ===\n")
+print("=== MSCMM Packing Script 2.1 ===\n")
 print("Collecting locale files...")
 
 BASE_DIR = os.path.dirname(os.getcwd())
 os.chdir(BASE_DIR + "\\OggConverter\\bin\\Release")
 
+
+def make_zip(files, zipName):
+    print('Creating new zip: {0}'.format(zipName))
+    NEW_ZIP = ZipFile(BASE_DIR + "\\" + zipName, 'w', zipfile.ZIP_DEFLATED)
+
+    for file in files:
+        NEW_ZIP.write(file)
+
+    NEW_ZIP.close()
+
+
+# Getting all locales
 LOCALES = []
 LOCALES = os.listdir("locales")
 
@@ -32,44 +42,31 @@ for locale in LOCALES:
     FILES.extend(["locales\\" + locale])
 
 FILES.extend(["MSC Music Manager.exe"])
-print("\nCreating mscmm_update.zip...")
-UPDATER_ZIP = ZipFile(BASE_DIR + "\\mscmm_update.zip",
-                      'w', zipfile.ZIP_DEFLATED)
-for file in FILES:
-    UPDATER_ZIP.write(file)
 
-UPDATER_ZIP.close()
+make_zip(FILES, 'mscmm_update.zip')
 
-print("mscmm_update.zip created succesfully!")
-print("Creating mscmm.zip...")
-
+# Adding ffmpeg and ffplay for full install
 FILES.extend(["ffmpeg.exe"])
 FILES.extend(["ffplay.exe"])
-FULL_ZIP = ZipFile(BASE_DIR + "\\mscmm.zip", "w", zipfile.ZIP_DEFLATED)
-for file in FILES:
-    FULL_ZIP.write(file)
 
-FULL_ZIP.close()
+make_zip(FILES, 'mscmm.zip')
 
-print("mscmm.zip created succesfully!")
-print("Removing logs and history files from Release...")
 
-if os.path.isfile("history.txt"):
-    os.remove("history.txt")
+def junk_cleaner(folder):
+    print('Now cleaning: {0}'.format(folder))
+    os.chdir(BASE_DIR + "\\OggConverter\\bin\\" + folder)
 
-if os.path.isdir("LOG"):
-    shutil.rmtree('LOG')
+    if os.path.isfile("history.txt"):
+        os.remove("history.txt")
 
-os.chdir(os.path.dirname(os.getcwd()))
-os.chdir("Debug")
-print("Removing logs and history files from Debug...")
+    if os.path.isdir("LOG"):
+        shutil.rmtree("LOG")
 
-if os.path.isfile("history.txt"):
-    os.remove("history.txt")
+    print('Junk cleaning of {0} done!'.format(folder))
 
-if os.path.isdir("LOG"):
-    shutil.rmtree('LOG')
 
-print("Done!")
+junk_cleaner('Release')
+junk_cleaner('Debug')
+
 print("Quitting...")
 quit()